+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/common/Readable.java
@@ -8,4 +8,5 @@ public interface Readable<T> {
     T read();
     Optional<T> readOpt();
     List<T> readAll();
+    long count();
 }